По-русски

+++ app/src/main/java/com/github/anniepank/hability/data/Habit.java
@@ -17,8 +17,8 @@ public class Habit {
     public final static LinkedHashMap<String, HabitType> namesAndImages = new LinkedHashMap<>();
     public boolean[] remindDays = {false, false, false, false, false, false, false};
     public boolean remind;
-    public long reminderHours;
-    public long reminderMinutes;
+    public int reminderHours;
+    public int reminderMinutes;
 
     static {
         namesAndImages.put("other", new HabitType("Other", R.drawable.other));
@@ -42,8 +42,8 @@ public class Habit {
         }
     }
 
-    public long getStreak() {
-        long streak = 0;
+    public int getStreak() {
+        int streak = 0;
         if (days.contains(DateUtilities.getToday())) {
             streak = 1;
         }
